requirements.txt
alembic==1.8.1
aniso8601==9.0.1
attrs==22.1.0
click==8.1.3
coverage==6.5.0
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
gunicorn==20.1.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.41
tomli==2.0.1
Werkzeug==2.1.2
zipp==3.8.1
